test: fix previous Home tests
